+++ v0.4/platform/ui/src/index.ts
@@ -15,7 +15,7 @@
 
 import { SvelteComponent } from 'svelte'
 
-import Root from './components/Root.svelte'
+import Root from './components/internal/Root.svelte'
 
 export type { AnyComponent } from './types'
 export { applicationShortcutKey } from './utils'
